Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.translate.nonfinite.worker-expected.txt


PASS translate() with Infinity/NaN is ignored
